push

+++ SwordOfUnityGame/Assets/Scripts/EnemyScripts/LogicScripts/AttackTypes/BuffOrDeBuff/Buff_Debuffs.cs
@@ -6,6 +6,6 @@ public abstract class Buff_Debuffs : AttackTypes
 {
     public enum AType { Buff, Debuff, Both}
     public byte RoundCounter;
-    public byte DamageMod;
+    public int DamageMod;
     public AType aType;
 }
