+++ SwordOfUnityGame/Assets/Scripts/ScriptableObjects/Items/WeaponItem.cs
@@ -2,6 +2,6 @@
 using System.Collections.Generic;
 using UnityEngine;
 public abstract class WeaponItem : ItemObjects {
-    public byte Damage;
+    public int Damage;
     public PlayerStatisticsManager pStatsMang;
 }
